Add router; organize components into separate files; make navbar transparent on home page only
