+++ src/App.css
@@ -1,7 +1,3 @@
 .content {
   background-color: #187be5;
 }
-
-.navbarstuff {
-  background: rgba(0, 120, 232, 0.6);
-}
