+++ src/app/Pages/TeamPage.js
@@ -1,0 +1,16 @@
+import 'bootstrap/dist/css/bootstrap.min.css';
+import React from 'react';
+import CommonNavbar from '../Components/CommonNavbar';
+
+export default function TeamPage() {
+  return (
+    <>
+      <CommonNavbar style={teamPageNavbarStyle} />
+      <div>not implemented</div>
+    </>
+  );
+}
+
+const teamPageNavbarStyle = {
+  background: "rgba(0, 120, 232, 0.6)",
+}
